+++ build.rs
@@ -63,7 +63,8 @@ fn check_device_feature() {
         std::process::exit(1);
     }
 
-    let device_variants: HashSet<String> = DEVICE_VARIANTS.iter().map(|s| s.to_string()).collect();
+    let device_variants: HashSet<String> =
+        DEVICE_VARIANTS.iter().map(|s| (*s).to_string()).collect();
 
     // get all selected features via env variables
     let selected_features: HashSet<String> = env::vars()
@@ -90,7 +91,7 @@ fn check_device_feature() {
         // pretty print all avaliable devices
         for line in device_variants {
             for device in line {
-                eprint!("{} ", device);
+                eprint!("{device} ");
             }
             eprintln!();
         }
@@ -185,24 +186,17 @@ This may be due to incorrect feature configuration in Cargo.toml or stm32f3xx-ha
     writeln!(file, "MEMORY {{").unwrap();
     writeln!(
         file,
-        "    FLASH (rx) : ORIGIN = 0x8000000, LENGTH = {}K",
-        flash
+        "    FLASH (rx) : ORIGIN = 0x8000000, LENGTH = {flash}K"
     )
     .unwrap();
     if ccmram > 0 {
         writeln!(
             file,
-            "    CCMRAM (rwx) : ORIGIN = 0x10000000, LENGTH = {}K",
-            ccmram
+            "    CCMRAM (rwx) : ORIGIN = 0x10000000, LENGTH = {ccmram}K"
         )
         .unwrap();
     }
-    writeln!(
-        file,
-        "    RAM (rwx) : ORIGIN = 0x20000000, LENGTH = {}K",
-        ram
-    )
-    .unwrap();
+    writeln!(file, "    RAM (rwx) : ORIGIN = 0x20000000, LENGTH = {ram}K").unwrap();
     writeln!(file, "}}").unwrap();
     println!("cargo:rustc-link-search={}", out_dir.display());
 }