+++ src/rcc/enable.rs
@@ -1,3 +1,4 @@
+#[allow(clippy::wildcard_imports)]
 use super::*;
 
 macro_rules! bus_enable {